Updating the UI via live data - reset works; the empty input error recognition fails
The implementation changes the initial value, so the original value is not equal to the or will never be equal to the choice items (other than the actual original reordering) since it is continuing from the retain state. So either this is the way to go or should figure out a way to set original item as retained value (which I doubt if could be properly synced as it is asynchronous and the first return is a empty list).