Fix#3146 : Create a generic utility for filtering enums  (#5529)
## Explanation

Fixes: #3146

This PR introduces a new utility function filterByEnumCondition to
standardize filtering of enums across various parts of the oppia-android
codebase. This utility function allows filtering of collections based on
a condition applied to enum values.

This PR introduces a new utility function filterByEnumCondition to
standardize filtering of enums across various parts of the oppia-android
codebase. This utility function allows filtering of collections based on
a condition applied to enum values.

**Key Changes**:
Added Utility Function:
filterByEnumCondition: A generic function to filter a collection based
on a condition applied to an enum extracted from each item in the
collection.

**Updated Existing Code**:
Refactored code in getLeastRecentMetricLogIndex,
getLeastRecentMediumPriorityEventIndex, and other methods to utilize the
new filterByEnumCondition function.
Updated the calculation of completedChapterCount and
inProgressChapterCount using the new utility function.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [ x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x ] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x ] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x ] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [ x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x ] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).
-

---------

Co-authored-by: Adhiambo Peres <59600948+adhiamboperes@users.noreply.github.com>