Using Transformations map to transform the live data
Though this works with live data for _choiceItems and rendering them with the latest wrong answer list, it fails to update the original choice items list and also throws an inital null value (and had a hot fix to it). This was not unit tested. May be trying out switch map could be an option with the asyncrhonous data.