Add NumberWithUnits + other fixes.
This commit adds support for NumberWithUnits, removes Kotlin reflection
support in Moshi, and makes some other loading improvements including:
- Better error handling when fetching versions (for faster failures).
- Cleaner null handling (fail on missing entry with no look-ahead object
  loading anymore; resulted in some new 'peekable' structures and
  missing properties being added).

This currently fails, however, due to some weird incompatibility between
a partially supported interaction and early loading failure. More
investigation and fixing is needed (this has been an issue that's
existed in the script for a long time).

Note also that this change is needed now due to the science classroom
heavily relying on number-with-units.