Fix ExplorationActivityTest and reverted ExplorationProgressModuleTest and FeatureFlagsLoggerTest back to TestPlatformParameterModule
Few test files use TestPlatformParameterModuleConstant values which was provided then by the TestPlatformParameterModule but those were just for testing purposes and not sure how to migrate them to production module since that would involve mixing both, either they need to be merged to be in one or handled as a separate file for constants just for testing and injected as required (also to be placed at a more appropriate location if required). Todo: 1. Migration of the left 6 files + these 2 -- 2. To remove TestPlatformParameterModule -- 3. To check for OppiaTestRule instance at every test file