Separating the class and method feature flag annotation extractions
The earlier implementation did 2 things 1. retrieved the annotations from both class and method levels and accumulated into one variable 2. it first sets the enable value followed by the disable value. To make sure the method annotation takes precedence the separation of class level and method level annotation retrievals are done and class levels and called first then goes the method levels, but still if there are multiple annotations for the same feature flag made this would always take the disable values for the final consideration which shouldn't be the case, so this need fix to just have one annotation for each class and method max 2 for a single feature flag / platform parameter; also this commit is just a temp test done with just the enable and disable feature flags which will need to be applied for platform parameters too (might need separate helper if gets too messy).