Huge amount of changes to the system.
This introduces module generation, feature flag & platform parameter
definitions, working implementations to the parameter system itself,
plus demonstrated and verified support for overrides in tests and in the
main app binary.

There are still a lot of broken tests that need fixing, and the app
itself is integrating the system incorrectly due to a very early
platform parameter injection requirement that was missed. The
initialization flow from the app startup through splash will need to be
redone a bit in order to hook up the new system correctly, but this
commit proves the system should be viable.