Revise TestFileCheck.
The new check performs much more thorough analysis to ensure that
exemptions point to valid files, and that directory structures match. It
also ensures that there are pairings for both tests to prod files and
vice versa.

The tests for this check have been completely rewritten, and a new
exemption type was added for extra test suites (such as local tests).

The exemptions proto had a lot of updates needed to remove invalid
references and to exempt these extra test suites.