Fix #5697: concept card close button (#5810)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
Fixes #5787, #5697
This PR includes fixing close button functionality of concept card.

### Root Cause & Fix
This issue is a regression stemming from an incorrect fragment manager
being used when calling dismissAll() in ConceptCardFragment.

In PR #5509 the dismiss functionality was implemented specifically for
the case when the concept card is shown from
**HintAndSolutionFragment**, where the correct fragment manager was
passed. However, the dismiss functionality for concept cards shown
directly from **StateFragment** was removed, leading to a missing
dismissal behavior in that scenario.

As a result, when a user closed the concept card that was launched by
**StateFragment**, the dismiss operation failed because there was no
proper handling or fragment manager provided for that case.

### Fixes

```
override fun dismissConceptCard() {
    getHintsAndSolution()?.dismissConceptCard()
        ?: explorationActivityPresenter.dismissConceptCard()
}
```
This fix first tries to delegate the dismissal to
**HintAndSolutionFragment**, ensuring the correct fragment manager is
used. If **HintAndSolutionFragment** is not present, it falls back to
the **ExplorationActivity's** presenter, which handles the case for
concept cards shown by **StateFragment**. This ensures that
**dismissAll**() is always called with the correct fragment manager,
resolving the issue for both cases.




## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [ ] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [ ] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [ ] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only

Closing concept card on StateFragment


https://github.com/user-attachments/assets/082b3d56-66c5-4f74-8c6e-e3e8af9fd462

Closing concept card on HintsAndSolutionDialogFragment


https://github.com/user-attachments/assets/70c3210b-0b0a-437a-a886-ec838f42aa4f

---------

Co-authored-by: Sneha Datta <snehadatta431@gmail.com>
Co-authored-by: Adhiambo Peres <59600948+adhiamboperes@users.noreply.github.com>
Co-authored-by: Mr. 17 <saptakmanna100@gmail.com>
Co-authored-by: RD Rama Devi <122200035+Rd4dev@users.noreply.github.com>
Co-authored-by: Ben Henning <ben@oppia.org>
Co-authored-by: Tobiloba Oyelekan <miketobi73@gmail.com>