Fix part of #59: Expand and sort module declarations in tests (#5824)
## Explanation

Fixes part of #59

Updates all test module declarations to use one line per module and sort
them.

This is a large change, but it's been 100% auto-generated using the
following code (run from within a script context):

```kotlin
fun main(vararg args: String) {
  // Path of the repo to be analyzed.
  val repoPath = "${args[0]}/"
  val testFiles = RepositoryFile.collectSearchFiles(
    repoPath = repoPath,
    expectedExtension = ".kt",
    exemptionsList = emptyList()
  ).filter { it.nameWithoutExtension.endsWith("Test") }
  testFiles.forEach { testFile ->
    val sourceContents = testFile.readText()
    val modulesPrefix = "modules = ["
    if (modulesPrefix in sourceContents) {
      val moduleNames =
        sourceContents
          .replace("\n", "")
          .replace(" ", "")
          .substringAfter("@Component(")
          .substringAfter("modules=[")
          .substringBefore("]")
          .split(",")
          .mapNotNull { it.takeIf(String::isNotEmpty) }
          .map { it.removeSuffix("::class") }
      val realModules = moduleNames.sorted().toSet()

      val fullyExpandedModules = realModules.joinToString(",\n      ") { "$it::class" }
      val linesToUse = if (realModules.isNotEmpty()) "  @Component(\n    modules = [\n      $fullyExpandedModules\n    ]\n  )" else "  @Component(modules = [])"

      val componentPrefix = "  @Component("
      val startIndex = sourceContents.indexOf(componentPrefix)
      val endIndex = sourceContents.substringAfter(componentPrefix).indexOf(')') + startIndex + componentPrefix.length
      val newSourceContents =
        "${sourceContents.substring(0, startIndex)}$linesToUse${sourceContents.substring(endIndex + 1)}"
      testFile.writeText(newSourceContents)
    }
  }
}
```

I am content with relying on the above generation (which is specifically
designed to avoid dropping components) and CI to ensure that this is
correct, though a spot check review would be helpful. There were two
classes `AppStartupStateControllerTest` and `DeprecationControllerTest`
that required some manual handling due to them both including line
comments within their module declarations, but these were easy to fix.

Note that this is being done now to simplify a bunch of downstream PRs
including #5823 and #5804 which are being done in turn to simplify #5725
(via another not-yet-created PR that will significantly reduce the
number of modules references that we need to maintain). This may expand
the number of lines of test code, but it's temporary until a few PRs
from now when they can be consolidated into new bundle modules (see
#5805).

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This is a purely infrastructural change.