Fix a LOT of tests.
This includes removing a lot of old platform parameter test code that
needs to be updated. Otherwise, it's primarily about using the test
platform parameter module in a bunch of places (since that automatically
initializes platform parameters and feature flags for prod code), which
in turn requires AssetModule (and both AssetModule and
TestPlatformParameterModule require explicit dependencies in Bazel).

There are still 31 known failing tests to fix, but the majority have
been addressed in this commit.