Fix part of #5734 and Fix #5897: Suppress Issue Categories and Include timer (#5921)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
<!--
- Explain what your PR does. If this PR fixes an existing bug, please
include
- "Fixes #bugnum:" in the explanation so that GitHub can auto-close the
issue
  - when this PR is merged.
  -->
Fixes part of #5734
Fix #5897 

This PR includes the following updates:
- Suppressed Issue categories that aren't relevant in lint analysis for
the codebase
- MissingTranslation: This is kept in check via TranslateWiki and can be
suppressed in Android lint reports
- SelectableText: It is preferred to keep the TextViews non-selectable
for now as being able to select the text would actively conflict with a
user action
- DuplicateStrings: some strings can get translated differently due to
the context within which it is being used, and keeping duplicated
strings is preferred and is a general false positive even in other
codebases' lint analysis
- SyntheticAccessors: The fix is really tedious since it involves
refactoring all the lambdas and suppressing them is preferred.
- GradleOverrides: This is a Gradle-related issue category and isn't
relevant since we have migrated to Bazel.
- Added a timer to display the script execution progress in the
terminal.
## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).