Fixes part of #5345: Fixes restarting of download on config change. (#5958)
Fixes part of #5345:
This pull request updates the logic for downloading remote platform
parameters in `PlatformParameterControllerDebugImpl`. The main change is
that logging for the download process now only occurs when a new
download is actually initiated, and not every time the method is called.

Improvements to download logic and logging:

* In `PlatformParameterControllerDebugImpl.kt`, changed the placement of
logging statements so that "Calling Force Download of remote parameters"
and "Download finished successfully." are logged only when a new
download task is started, preventing duplicate logs when a download is
already in progress.
Added a check :
```kotlin
   if (ongoingDownloadTask == null || ongoingDownloadTask?.isCancelled == true) {
```
To simulate actual download behaviour and verify that on config change
download doesn't restarts please repalace:
```
 // TODO(#5950): Replace with actual logic to force-download remote parameters.
```
with

```kotlin
delay(3000)
```
#### Evaluation Points:

1. Rotating the device while the downloads dialog is open and actively
downloading does not causes it to restart the call from scratch.
[Video](https://github.com/user-attachments/assets/40aa156d-240f-4ae7-85c9-8538db102700)
2. Rotating the device when the ‘successfully downloaded’ dialog is open
does not causes the download to start again.

[Video](https://github.com/user-attachments/assets/728c30b2-16c8-476b-9063-7571b63f96c4)



</div></b>

#### Screenrecording:

##### Before

[Screencast from 2025-09-09
14-46-31.webm](https://github.com/user-attachments/assets/40aa156d-240f-4ae7-85c9-8538db102700)

##### After

[Screencast from 2025-09-09
09-42-16.webm](https://github.com/user-attachments/assets/95c4dd7a-6098-453f-a94e-6a4ebdb4afe8)



## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).