Fix #5736, Fix #5721, Fix 5954, Fix 5636 and Fix part of #5800: 0.16 beta bug fixes (#5831)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation

Fix #5736: The last character in the list is truncated because

https://github.com/oppia/oppia-android/blob/cda4f44e36222fae5fc8672c774cc8c281f97156/utility/src/main/java/org/oppia/android/util/parser/html/HtmlParser.kt#L179
deletes 2 characters from the end if the text ends with a single `\n`,
which is too aggressive. This means it deletes the last character (4)
and the \n, when it should only delete the \n.

The safest solution is to trim all leading and trailing `\n` characters,
not just one. That way, whether there’s 1, 2, or 10 newlines at the
edges, they’ll be removed without touching non-\n characters.

|Before|After|
|--|--|

|![Image](https://github.com/user-attachments/assets/c43a8089-ff53-486f-94d6-0fc07ae9b37b)|![image](https://github.com/user-attachments/assets/7cca97e7-9496-44da-9333-854954975fd9)|

Fix https://github.com/oppia/oppia-android/issues/5721:

Per @subhajitxyz's proposed solution in
https://github.com/oppia/oppia-android/issues/5721#issuecomment-2854101478,
We need to calculate the width of the text outside of
`drawLeadingMargin()`, so that `getLeadingMargin()` can return the
correct computedLeadingMargin independently. We do this by passing a
reference of the TextView to the tag handler so that we can use it to
construct a matching `TextPaint` object and dynamically calculate the
prefix width (e.g., "1.", "10.") using its font, text size, and other
styling. This will allow us to decouple `getLeadingMargin()` from
`drawLeadingMargin()` and compute the margin correctly during layout.

|||
|--|--|

|![Image](https://github.com/user-attachments/assets/f0e98f58-301c-4eef-898d-93062604baf1)|
![Screenshot_2025-09-04-23-43-44-057_org oppia
android](https://github.com/user-attachments/assets/a9c4e003-c494-4302-842c-b9edf1e52144)|

|![Image](https://github.com/user-attachments/assets/7c4682a2-2542-403f-aeaa-c2cc1a87765a)|![Screenshot_2025-09-04-23-40-19-274_org
oppia
android](https://github.com/user-attachments/assets/20453581-77de-4201-8944-c91acd23c61a)|

Fix #5636

The developer options and admin controls menu item faailed to reopen
because of respective variables in `NavigationDrawerFragmentPresenter`
being set to true to indicate that they are already open, as well as
forcing the drawer to close.

Unfortunately, tests for repeated opening are blocked by known issue:
https://github.com/oppia/oppia-android/issues/2535.

Fix #5954

This bug is caused by the survey being shown AFTER the exploration has
been stopped, so attempting to play the exploration results into the
frozen state. The user is still able to exit the exploration normally by
clicking the close or return to topic buttons again, but this creates an
inconsistent UX.

The bug is fixed by navigating the user away from the
ExplorationActivity if they decline the survey.

Some minor chnages have been made to the survey code to remove some
redundancies.

|Before|After|
|--|--|

|https://github.com/user-attachments/assets/5bed7d7b-6d78-427a-a105-86fc0f15fd08|https://github.com/user-attachments/assets/cb6f042f-3e2c-4ae7-b972-75cf08a2f137|

Fix part of #5800

The survey gets dismissed due to the gating provider updating as a
result of the combined profile DataProvider updates. Immediately the
survey popup is show, the `surveyLastShownTimestampMs` field of a
profile gets update, and this updates downstream providers that are
subscribed to the profile provider. This has been fixed by removing the
observer to the gating provider once it has returned the first
time.vThis happened only on clicking "return to question".

This was not possible to test in `StateFragmentTest` because of issues
with dragAndDrop on robolectric, preventing playing through and
exploration. I spent a considerable amount of time trying to set up the
tst using the Ratios exploration, but experienced difficulty with item
selection in state 19 -- the images are large and scrolling is needed to
see and click all the options, but I was not able to make this work at
this time. So keeping the issue open to track adding tests either buy
fixing the ratios playthrough, or the drag and drop issue.

|Before|After|
|--|--|
|.
|https://github.com/user-attachments/assets/cb6f042f-3e2c-4ae7-b972-75cf08a2f137|

The PR also bumps versions for the release.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).